AXFS: fix type cast warning in axfs_region_is_vmalloc()
This patch fixes following compilation warning caused by
Commit-ID: a7e526129f6ea298513ef4ce7aa6db08fc069d14

fs/axfs/axfs_inode.c: In function axfs_region_is_vmalloc:
fs/axfs/axfs_inode.c:175:11: warning: cast from pointer to integer of
different size [-Wpointer-to-int-cast]

+++ fs/axfs/axfs_inode.c
@@ -172,7 +172,7 @@ static inline u64 axfs_get_xip_region_physaddr(struct axfs_super *sbi)
 static inline int axfs_region_is_vmalloc(struct axfs_super *sbi,
 					 struct axfs_region_desc *region)
 {
-	u64 va = (u32) region->virt_addr;
+	u64 va = (unsigned long) region->virt_addr;
 	u64 vo = (u64) region->fsoffset + (u64) sbi->virt_start_addr;
 
 	if (va == 0)
